requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
flask==2.3.3
aiohttp==3.9.3
requests
beautifulsoup4
pygame
aiohttp==3.9.3
beautifulsoup4
discord.py==2.3.2
flask==2.3.3
pygame
python-dotenv==1.0.0
requests